Use int64 for limit/offset values to ensure values greater than max 32-bit integer are addressable.

+++ dialect_common.go
@@ -122,7 +122,7 @@ func (s commonDialect) currentDatabase() (name string) {
 	return
 }
 
-func (commonDialect) LimitAndOffsetSQL(limit, offset int) (sql string) {
+func (commonDialect) LimitAndOffsetSQL(limit, offset int64) (sql string) {
 	if limit > 0 || offset > 0 {
 		if limit >= 0 {
 			sql += fmt.Sprintf(" LIMIT %d", limit)
